requirements.txt
python-dotenv

flask
flask-socketio
apscheduler


pillow

torch 
torchvision
torchaudio --index-url https://download.pytorch.org/whl/cu121
ultralytics
scikit-learn